% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_gmt}
\alias{load_gmt}
\title{Load a GMT (Gene Matrix Transposed) file into a [`VennDataset-class`]}
\usage{
load_gmt(path)
}
\arguments{
\item{path}{Path to the .gmt file.}
}
\value{
A [`VennDataset-class`].
}
\description{
Each line is one set: `set_name<TAB>description<TAB>item1<TAB>item2<TAB>...`.
Lines with fewer than 3 tab-separated columns or empty set names are skipped.
}
\examples{
tmp <- tempfile(fileext = ".gmt")
writeLines(c("SetA\tdesc\tGENE1\tGENE2\tGENE3",
             "SetB\tdesc\tGENE2\tGENE3\tGENE4"), tmp)
ds <- load_gmt(tmp)
ds@set_names
}
